6/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
passlib[bcrypt]
python-dotenv
python-jose
